ISA 2.0 migration: codebase compiles, doesn't work - need to add BRANCH instructions to the end of each BasicBlock

+++ ir/typedefs.h
@@ -5,6 +5,6 @@
 #include <cstdint>
 
 typedef size_t IdType;
-typedef int ImmType;
+typedef double ImmType;
 
 #endif
